+++ backend/inventorysimulator/src/main/java/com/retailops/inventorysimulator/simulator/dto/NewsVendorResponse.java
@@ -1,8 +1,9 @@
 package com.retailops.inventorysimulator.simulator.dto;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public record NewsVendorResponse(String product,
                                  BigDecimal criticalRatio,
-                                 int optimalOrderQuantity) {
+                                 BigInteger optimalOrderQuantity) {
 }
